+++ tests/test_force_reload.py
@@ -9,6 +9,11 @@ import ipytest
         ("foo", ("foo", "foo.bar", "foo.baz")),
         ("foo.bar", ("foo.bar",)),
         ("f", ()),
+        ("f*", ("foo", "foo.bar", "foo.baz")),
+        ("test_*", ("test_example1", "test_example2")),
+        pytest.param("*.", (), id="invalid pattern"),
+        ("*.bar", ("foo.bar",)),
+        ("foo.ba?", ("foo.bar", "foo.baz")),
     ],
 )
 def test_force_reload_example(include, removed):
@@ -16,10 +21,13 @@ def test_force_reload_example(include, removed):
         "foo": None,
         "foo.bar": None,
         "foo.baz": None,
+        "test_example1": None,
+        "test_example2": None,
     }
+    expected = set(modules) - set(removed)
 
     ipytest.force_reload(include, modules=modules)
-    assert set(modules) == (set(modules) - set(removed))
+    assert set(modules) == expected
 
 
 def test_forc_reload():